java_jni/refactoring: move class and interface related code to separate files
