+++ macroslib/src/java_jni/java_code.rs
@@ -1,26 +1,10 @@
-use std::{io::Write, path::Path};
-
 use bitflags::bitflags;
-use rustc_hash::FxHashSet;
-use smol_str::SmolStr;
 
-use super::{
-    map_write_err, method_name, JavaConverter, JavaForeignTypeInfo, JniForeignMethodSignature,
-    NullAnnotation,
-};
-use crate::{
-    file_cache::FileWriteCache,
-    namegen::new_unique_name,
-    typemap::{
-        ast::if_result_return_ok_err_types, TypeMap, FROM_VAR_TEMPLATE, TO_VAR_TEMPLATE,
-        TO_VAR_TYPE_TEMPLATE,
-    },
-    types::{ForeignInterface, ForeignerClassInfo, MethodAccess, MethodVariant},
-    WRITE_TO_MEM_FAILED_MSG,
-};
+use super::{JniForeignMethodSignature, NullAnnotation};
+use crate::WRITE_TO_MEM_FAILED_MSG;
 
 bitflags! {
-    struct ArgsFormatFlags: u8 {
+    pub(in crate::java_jni) struct ArgsFormatFlags: u8 {
         const NONE = 0;
         const USE_COMMA_IF_NEED = 1;
         const EXTERNAL = 2;
@@ -29,474 +13,7 @@ bitflags! {
     }
 }
 
-pub(in crate::java_jni) fn generate_java_code_for_interface(
-    output_dir: &Path,
-    package_name: &str,
-    interface: &ForeignInterface,
-    methods_sign: &[JniForeignMethodSignature],
-    use_null_annotation: Option<&str>,
-) -> Result<(), String> {
-    let path = output_dir.join(format!("{}.java", interface.name));
-    let mut file = FileWriteCache::new(&path);
-    let imports = get_null_annotation_imports(use_null_annotation, methods_sign);
-    let interface_comments = doc_comments_to_java_comments(&interface.doc_comments, true);
-    write!(
-        file,
-        r#"// Automaticaly generated by rust_swig
-package {package_name};
-{imports}
-{doc_comments}
-public interface {interface_name} {{
-"#,
-        package_name = package_name,
-        interface_name = interface.name,
-        doc_comments = interface_comments,
-        imports = imports,
-    )
-    .map_err(&map_write_err)?;
-
-    for (method, f_method) in interface.items.iter().zip(methods_sign) {
-        write!(
-            file,
-            r#"
-{doc_comments}
-    void {method_name}({single_args_with_types});
-"#,
-            method_name = method.name,
-            doc_comments = doc_comments_to_java_comments(&method.doc_comments, false),
-            single_args_with_types = args_with_java_types(
-                f_method,
-                method.arg_names_without_self(),
-                ArgsFormatFlags::EXTERNAL,
-                use_null_annotation.is_some()
-            ),
-        )
-        .expect(WRITE_TO_MEM_FAILED_MSG);
-    }
-
-    write!(
-        file,
-        r#"
-}}
-"#,
-    )
-    .expect(WRITE_TO_MEM_FAILED_MSG);
-    file.update_file_if_necessary().map_err(&map_write_err)?;
-    Ok(())
-}
-
-pub(in crate::java_jni) fn generate_java_code(
-    conv_map: &mut TypeMap,
-    output_dir: &Path,
-    package_name: &str,
-    class: &ForeignerClassInfo,
-    methods_sign: &[JniForeignMethodSignature],
-    null_annotation_package: Option<&str>,
-) -> Result<(), String> {
-    let java_rust_self_name = "mNativeObj";
-
-    let path = output_dir.join(format!("{}.java", class.name));
-    let mut file = FileWriteCache::new(&path);
-
-    let imports = get_null_annotation_imports(null_annotation_package, methods_sign);
-
-    let class_doc_comments = doc_comments_to_java_comments(&class.doc_comments, true);
-    write!(
-        file,
-        r#"// Automaticaly generated by rust_swig
-package {package_name};
-{imports}
-{doc_comments}
-public final class {class_name} {{
-"#,
-        package_name = package_name,
-        imports = imports,
-        class_name = class.name,
-        doc_comments = class_doc_comments,
-    )
-    .expect(WRITE_TO_MEM_FAILED_MSG);
-
-    let mut have_methods = false;
-    let mut have_constructor = false;
-
-    for (method, f_method) in class.methods.iter().zip(methods_sign) {
-        write!(
-            &mut file,
-            "{doc_comments}",
-            doc_comments = doc_comments_to_java_comments(&method.doc_comments, false)
-        )
-        .expect(WRITE_TO_MEM_FAILED_MSG);
-
-        let may_return_error = match method.fn_decl.output {
-            syn::ReturnType::Default => false,
-            syn::ReturnType::Type(_, ref ptype) => {
-                let ret_rust_ty = conv_map.find_or_alloc_rust_type(ptype, class.src_id);
-                if_result_return_ok_err_types(&ret_rust_ty).is_some()
-            }
-        };
-
-        let exception_spec = if may_return_error {
-            " throws Exception"
-        } else {
-            ""
-        };
-
-        let method_access = match method.access {
-            MethodAccess::Private => "private",
-            MethodAccess::Public => "public",
-            MethodAccess::Protected => unreachable!(),
-        };
-        let conv_code_flags = match method.variant {
-            MethodVariant::StaticMethod => ArgsFormatFlags::INTERNAL,
-            MethodVariant::Method(_) => ArgsFormatFlags::COMMA_BEFORE | ArgsFormatFlags::INTERNAL,
-            MethodVariant::Constructor => ArgsFormatFlags::INTERNAL,
-        };
-        let mut known_names: FxHashSet<SmolStr> =
-            method.arg_names_without_self().map(|x| x.into()).collect();
-        if let MethodVariant::Method(_) = method.variant {
-            if known_names.contains(java_rust_self_name) {
-                return Err(format!("In method {} there is argument with name {}, this name reserved for generated code",
-                                   method.short_name(), java_rust_self_name));
-            }
-            known_names.insert(java_rust_self_name.into());
-        }
-        let ret_name = new_unique_name(&known_names, "ret");
-        known_names.insert(ret_name.clone());
-        let conv_ret = new_unique_name(&known_names, "conv_ret");
-        known_names.insert(conv_ret.clone());
-
-        let (convert_code, args_for_call_internal) = convert_code_for_method(
-            f_method,
-            method.arg_names_without_self(),
-            known_names,
-            conv_code_flags,
-        );
-        let func_name = method_name(method, f_method);
-
-        let external_args_except_self = args_with_java_types(
-            f_method,
-            method.arg_names_without_self(),
-            ArgsFormatFlags::EXTERNAL,
-            null_annotation_package.is_some(),
-        );
-        fn calc_output_conv<'a>(
-            output: &'a JavaForeignTypeInfo,
-            conv: &'a JavaConverter,
-            ret_name: &str,
-            conv_ret: &str,
-        ) -> (&'a str, &'a str, String) {
-            let ret_type = output.base.name.as_str();
-            let intermidiate_ret_type = conv.java_transition_type.as_str();
-            let conv_code = conv
-                .converter
-                .replace(FROM_VAR_TEMPLATE, ret_name)
-                .replace(TO_VAR_TYPE_TEMPLATE, &format!("{} {}", ret_type, conv_ret))
-                .replace(TO_VAR_TEMPLATE, &conv_ret)
-                .replace("@NonNull", "")
-                .replace("@Nullable", "")
-                .trim()
-                .into();
-            (ret_type, intermidiate_ret_type, conv_code)
-        }
-
-        let (ret_type, intermidiate_ret_type, ret_conv_code) = match method.variant {
-            MethodVariant::StaticMethod => {
-                if let Some(conv) = f_method.output.java_converter.as_ref() {
-                    calc_output_conv(&f_method.output, conv, &ret_name, &conv_ret)
-                } else {
-                    let ret_type = f_method.output.base.name.as_str();
-                    (ret_type, ret_type, String::new())
-                }
-            }
-            MethodVariant::Method(_) => {
-                if let Some(conv) = f_method.output.java_converter.as_ref() {
-                    calc_output_conv(&f_method.output, conv, &ret_name, &conv_ret)
-                } else {
-                    let ret_type = f_method.output.base.name.as_str();
-                    (ret_type, ret_type, String::new())
-                }
-            }
-            MethodVariant::Constructor => ("long", "long", String::new()),
-        };
-
-        let need_conversation = !convert_code.is_empty() || !ret_conv_code.is_empty();
-
-        match method.variant {
-            MethodVariant::StaticMethod => {
-                let (native, end) = if !need_conversation {
-                    ("native ", ";\n")
-                } else {
-                    ("", " {\n")
-                };
-                write!(
-                    file,
-                    r#"
-    {method_access} static {native}{ret_type} {func_name}({args_with_types}){exception_spec}{end}"#,
-                    method_access = method_access,
-                    ret_type = ret_type,
-                    func_name = method.short_name(),
-                    args_with_types = external_args_except_self,
-                    exception_spec = exception_spec,
-                    native = native,
-                    end = end,
-                )
-                .expect(WRITE_TO_MEM_FAILED_MSG);
-
-                if need_conversation {
-                    if !convert_code.is_empty() {
-                        let mut code = convert_code.as_bytes();
-                        if code[0] == b'\n' {
-                            code = &code[1..];
-                        }
-                        file.write_all(code).expect(WRITE_TO_MEM_FAILED_MSG);
-                        file.write_all(b"\n").expect(WRITE_TO_MEM_FAILED_MSG);
-                    }
-                    if ret_conv_code.is_empty() {
-                        write!(
-                            file,
-                            r#"        {return_code}{func_name}({args});
-    }}"#,
-                            func_name = func_name,
-                            return_code = if ret_type != "void" { "return " } else { "" },
-                            args = args_for_call_internal,
-                        )
-                        .expect(WRITE_TO_MEM_FAILED_MSG);
-                    } else {
-                        write!(
-                            file,
-                            r#"        {intermidiate_ret_type} {ret_name} = {func_name}({args});
-        {ret_conv_code}
-"#,
-                            ret_conv_code = ret_conv_code,
-                            ret_name = ret_name,
-                            intermidiate_ret_type = intermidiate_ret_type,
-                            func_name = func_name,
-                            args = args_for_call_internal,
-                        )
-                        .expect(WRITE_TO_MEM_FAILED_MSG);
-                        if ret_type != "void" {
-                            write!(
-                                file,
-                                r#"
-        return {conv_ret};
-"#,
-                                conv_ret = conv_ret
-                            )
-                            .expect(WRITE_TO_MEM_FAILED_MSG);
-                        }
-                        file.write_all("    }".as_bytes())
-                            .expect(WRITE_TO_MEM_FAILED_MSG);
-                    }
-
-                    write!(
-                        file,
-                        r#"
-    private static native {intermidiate_ret_type} {func_name}({args_with_types}){exception_spec};
-                           "#,
-                        func_name = func_name,
-                        intermidiate_ret_type = intermidiate_ret_type,
-                        exception_spec = exception_spec,
-                        args_with_types = args_with_java_types(
-                            f_method,
-                            method.arg_names_without_self(),
-                            ArgsFormatFlags::INTERNAL,
-                            null_annotation_package.is_some()
-                        ),
-                    )
-                    .expect(WRITE_TO_MEM_FAILED_MSG);
-                }
-            }
-            MethodVariant::Method(_) => {
-                have_methods = true;
-                write!(
-                    file,
-                    r#"
-    {method_access} final {ret_type} {method_name}({single_args_with_types}){exception_spec} {{"#,
-                    method_access = method_access,
-                    ret_type = ret_type,
-                    method_name = method.short_name(),
-                    exception_spec = exception_spec,
-                    single_args_with_types = external_args_except_self,
-                )
-                .expect(WRITE_TO_MEM_FAILED_MSG);
-                if !convert_code.is_empty() {
-                    if convert_code.as_bytes()[0] != b'\n' {
-                        file.write_all(b"\n").expect(WRITE_TO_MEM_FAILED_MSG);
-                    }
-                    file.write_all(convert_code.as_bytes())
-                        .expect(WRITE_TO_MEM_FAILED_MSG);
-                }
-                if ret_conv_code.is_empty() {
-                    write!(
-                        file,
-                        r#"
-        {return_code}{func_name}({rust_self_name}{args});
-"#,
-                        rust_self_name = java_rust_self_name,
-                        return_code = if ret_type != "void" { "return " } else { "" },
-                        args = args_for_call_internal,
-                        func_name = func_name,
-                    )
-                    .expect(WRITE_TO_MEM_FAILED_MSG);
-                } else {
-                    write!(
-                        file,
-                        r#"
-        {intermidiate_ret_type} {ret_name} = {func_name}({rust_self_name}{args});
-        {ret_conv_code}
-"#,
-                        rust_self_name = java_rust_self_name,
-                        ret_conv_code = ret_conv_code,
-                        ret_name = ret_name,
-                        intermidiate_ret_type = intermidiate_ret_type,
-                        func_name = func_name,
-                        args = args_for_call_internal,
-                    )
-                    .expect(WRITE_TO_MEM_FAILED_MSG);
-                    if ret_type != "void" {
-                        write!(
-                            file,
-                            r#"
-        return {conv_ret};
-"#,
-                            conv_ret = conv_ret
-                        )
-                        .expect(WRITE_TO_MEM_FAILED_MSG);
-                    }
-                }
-
-                file.write_all("    }".as_bytes())
-                    .expect(WRITE_TO_MEM_FAILED_MSG);
-
-                write!(
-                    file,
-                    r#"
-    private static native {intermidiate_ret_type} {func_name}(long self{args_with_types}){exception_spec};
-"#,
-                    intermidiate_ret_type = intermidiate_ret_type,
-                    exception_spec = exception_spec,
-                    func_name = func_name,
-                    args_with_types = args_with_java_types(
-                        f_method,
-                        method.arg_names_without_self(),
-                        ArgsFormatFlags::USE_COMMA_IF_NEED | ArgsFormatFlags::INTERNAL,
-                        null_annotation_package.is_some()
-                    ),
-                )
-                .expect(WRITE_TO_MEM_FAILED_MSG);
-            }
-            MethodVariant::Constructor => {
-                have_constructor = true;
-
-                if method.is_dummy_constructor() {
-                    write!(
-                        file,
-                        "
-    {method_access} {class_name}() {{}}
-",
-                        method_access = method_access,
-                        class_name = class.name,
-                    )
-                    .expect(WRITE_TO_MEM_FAILED_MSG);
-                } else {
-                    write!(
-                        file,
-                        r#"
-    {method_access} {class_name}({ext_args_with_types}){exception_spec} {{
-"#,
-                        method_access = method_access,
-                        class_name = class.name,
-                        exception_spec = exception_spec,
-                        ext_args_with_types = external_args_except_self,
-                    )
-                    .expect(WRITE_TO_MEM_FAILED_MSG);
-                    if !convert_code.is_empty() {
-                        let mut code = convert_code.as_bytes();
-                        if code[0] == b'\n' {
-                            code = &code[1..];
-                        }
-                        file.write_all(code).expect(WRITE_TO_MEM_FAILED_MSG);
-                        file.write_all(b"\n").expect(WRITE_TO_MEM_FAILED_MSG);
-                    }
-                    write!(
-                        file,
-                        "        {rust_self_name} = init({args});
-    }}
-    private static native long {func_name}({args_with_types}){exception_spec};
-",
-                        rust_self_name = java_rust_self_name,
-                        exception_spec = exception_spec,
-                        func_name = func_name,
-                        args_with_types = args_with_java_types(
-                            f_method,
-                            method.arg_names_without_self(),
-                            ArgsFormatFlags::INTERNAL,
-                            null_annotation_package.is_some()
-                        ),
-                        args = args_for_call_internal,
-                    )
-                    .expect(WRITE_TO_MEM_FAILED_MSG);
-                }
-            }
-        }
-    }
-
-    if have_methods && !have_constructor {
-        return Err(format!(
-            "package {}, class {}: has methods, but no constructor\n
-May be you need to use `private constructor = empty;` syntax?",
-            package_name, class.name
-        ));
-    }
-    if have_constructor {
-        write!(
-            file,
-            "
-    public synchronized void delete() {{
-        if ({rust_self_name} != 0) {{
-            do_delete({rust_self_name});
-            {rust_self_name} = 0;
-       }}
-    }}
-    @Override
-    protected void finalize() throws Throwable {{
-        try {{
-            delete();
-        }}
-        finally {{
-             super.finalize();
-        }}
-    }}
-    private static native void do_delete(long me);
-    /*package*/ long {rust_self_name};
-",
-            rust_self_name = java_rust_self_name,
-        )
-        .expect(WRITE_TO_MEM_FAILED_MSG);
-    }
-
-    //utility class, so add private constructor
-    //to prevent object creation
-    if !have_constructor && !have_methods {
-        write!(
-            file,
-            r#"
-    private {class_name}() {{}}
-"#,
-            class_name = class.name
-        )
-        .map_err(&map_write_err)?;
-    }
-
-    file.write_all(class.foreigner_code.as_bytes())
-        .expect(WRITE_TO_MEM_FAILED_MSG);
-    write!(file, "}}").expect(WRITE_TO_MEM_FAILED_MSG);
-
-    file.update_file_if_necessary().map_err(&map_write_err)?;
-    Ok(())
-}
-
-fn args_with_java_types<'a, NI: Iterator<Item = &'a str>>(
+pub(in crate::java_jni) fn args_with_java_types<'a, NI: Iterator<Item = &'a str>>(
     method: &JniForeignMethodSignature,
     arg_name_iter: NI,
     flags: ArgsFormatFlags,
@@ -534,45 +51,6 @@ fn args_with_java_types<'a, NI: Iterator<Item = &'a str>>(
     res
 }
 
-fn convert_code_for_method<'a, NI: Iterator<Item = &'a str>>(
-    f_method: &JniForeignMethodSignature,
-    arg_name_iter: NI,
-    mut known_names: FxHashSet<SmolStr>,
-    flags: ArgsFormatFlags,
-) -> (String, String) {
-    let mut conv_code = String::new();
-    let mut args_for_call_internal = String::new();
-
-    if flags.contains(ArgsFormatFlags::COMMA_BEFORE) && !f_method.input.is_empty() {
-        args_for_call_internal.push_str(", ");
-    }
-
-    for (i, (arg, arg_name)) in f_method.input.iter().zip(arg_name_iter).enumerate() {
-        let after_conv_arg_name =
-            if let Some(code) = arg.java_converter.as_ref().map(|x| &x.converter) {
-                let templ = format!("a{}", i);
-                let after_conv_arg_name = new_unique_name(&known_names, &templ);
-                known_names.insert(after_conv_arg_name.clone());
-                let java_code = code
-                    .replace(TO_VAR_TEMPLATE, &after_conv_arg_name)
-                    .replace(FROM_VAR_TEMPLATE, arg_name);
-                conv_code.push_str(&java_code);
-                Some(after_conv_arg_name)
-            } else {
-                None
-            };
-        if let Some(after_conv_arg_name) = after_conv_arg_name {
-            args_for_call_internal.push_str(&after_conv_arg_name);
-        } else {
-            args_for_call_internal.push_str(arg_name);
-        }
-        if i != (f_method.input.len() - 1) {
-            args_for_call_internal.push_str(", ");
-        }
-    }
-    (conv_code, args_for_call_internal)
-}
-
 pub(in crate::java_jni) fn doc_comments_to_java_comments(
     doc_comments: &[String],
     class_comments: bool,
@@ -606,7 +84,7 @@ pub(in crate::java_jni) fn doc_comments_to_java_comments(
     comments
 }
 
-fn get_null_annotation_imports(
+pub(in crate::java_jni) fn get_null_annotation_imports(
     null_annotation_package: Option<&str>,
     methods_sign: &[JniForeignMethodSignature],
 ) -> String {
